updated SHA256 checksum how-to
It is now possible to check the integrity of the Mainnet-Scripts via SHA256 checksums. In addition to the Github hosted verification file it is using the verification file from an additional external secure webserver. Also you can check with this method if you're running the latest version of the scripts.